IteratorSerializer removed as it's not needed. Existing tests don't throw any concurrent modification exception
